fix: 일지등록 페이지 컴포넌트 분리
